Antifraude/requirements.txt
kafka-python
sqlalchemy
pymysql
scikit-learn
pandas
